+++ CHANGELOG.md
@@ -12,3 +12,5 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changes
 - Minimum Supported Shopware version is now 6.4.5.0
 - Add new monolog channel "mollie". You can now add custom handler and assign them to the mollie channel
+### Fixes
+- fixed order details in refund manager in shopware 6.4.x